notebooks/week2_openai/deep_research/requirements.txt
# OpenAI Agents SDK
openai-agents>=0.1.0

# Environment management
python-dotenv>=1.0.0

# Web interface
gradio>=4.0.0

# HTTP requests for email
requests>=2.31.0

# Data handling
pydantic>=2.0.0

# Async support
asyncio 